refactor & add api service
