+++ client/MusiConnect/App.js
@@ -1,31 +1,7 @@
-import { StatusBar } from 'expo-status-bar';
-import React, {useState} from 'react';
-import { Button, StyleSheet, Text, View } from 'react-native';
-
+import React from 'react';
+import HomeScreen from './containers/Home/HomeScreen'
 export default function App() {
-  const [num, setNum] = useState(1)
-  const func = () => {
-    console.log("PRESSED BUTTON")
-    setNum(num+1)
-  }
-  return (
-    <View style={styles.container}>
-      <Text style={styles.h1}>Music Connect!</Text>
-      <Text>Number: {num}</Text>
-      <Button title="Increment" onPress={func}></Button>
-      <StatusBar style="auto" />
-    </View>
-  );
+  return(
+    <HomeScreen />
+  )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  h1 : {
-    fontSize: 50
-  }
-});
