+++ client/MusiConnect/Consts.js
@@ -1,0 +1,1 @@
+export const SERVER_URL = 'https://api.song.link/v1-alpha.1/links'
